requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
flake8==4.0.1
idna==3.3
lxml==4.8.0
mccabe==0.6.1
numpy==1.22.3
pandas==1.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.9
